Fix Powerline glyphs
[why]
The Powerline glyphs protrude noticeable down into the next line.

The reason is that the lowest point of that glyphs is at
-340 while the descenders is only set to -250. The ascenders and
descenders is used by terminals to determine baseline to baseline
distance (together with the gap, sometimes, which is zero in this
font).

[how]
Move the bottom and middle points of the Powerline glyphs
E0B0 E0B1 E0B2 E0B3

The top point has an overshoot of 10 (line top is at 1100, top point
coordinate is at 1110), so kept the same amount of overshoot for the
bottom point: line bottom is at -250, bottom point should be a -260.

Bottom:
-340 -> -260
-291 -> -211    (same distance up)

Middle:
 385 ->  425    (half the distance to keep centered)

[note]
Did the changes in Glyph3 but commited only the basic changes while
keeping the version number in the glyphs file.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>